+++ peppol-icd/src/main/java/no/difi/vefa/peppol/icd/api/Icd.java
@@ -29,4 +29,6 @@ public interface Icd {
 
     Scheme getScheme();
 
+    String getIssuingAgency();
+
 }